+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,3 +1,3 @@
 public interface IDamageable {
-    void GetHit(int damageSuffered);
+    void GetHit(float damageSuffered);
 }
